Extract CardefPool to pool.ts
