+++ src/cards.ts
@@ -1,6 +1,4 @@
-import { CardId, CardNumber, CardType, SetId } from './types';
-import { readFileSync } from 'fs';
-import { resolve } from 'path';
+import { CardNumber, CardType, SetId } from './types';
 
 export interface Cardef {
     setId: SetId;
@@ -14,7 +12,7 @@ export interface Cardef {
     // rarity
 }
 
-function createCardef(
+export function createCardef(
     setId?: SetId,
     cardNumber?: CardNumber,
     name?: string,
@@ -36,77 +34,4 @@ function createCardef(
         power,
         vp: validVp,
     };
-}
-
-export class CardefPool {
-    private pool: Map<CardId, Cardef>;
-
-    public constructor() {
-        this.pool = new Map<CardId, Cardef>();
-        const path = resolve(__dirname, '../resources/cardpool.tsv');
-        this.loadFromFile(path);
-        if (this.pool.size !== 100) {
-            throw new Error(`Pool not 100 cards! (${this.pool.size})`);
-        }
-    }
-
-    public loadFromTsvString(tsvString: string): void {
-        const rows = tsvString.split('\n');
-        rows.shift();
-        const validRows = rows.filter((row) => {
-            return row.length > 0;
-        });
-        const cardefs = validRows.map((row) => {
-            const columns = row.split('\t');
-            const setId = this.forceToString(columns.shift());
-            const id = columns.shift();
-            const type = columns.shift();
-            const name = columns.shift();
-            const power = parseInt(columns.shift() ?? '');
-            const vp = parseInt(columns.shift() ?? '999');
-
-            try {
-                return createCardef(
-                    setId,
-                    id,
-                    name,
-                    type as CardType,
-                    power,
-                    vp
-                );
-            } catch (e: unknown) {
-                console.log(e);
-                console.log(row);
-                throw e;
-            }
-        });
-
-        this.pool.clear();
-        cardefs.forEach((cardef) => {
-            const fullId = `${cardef.setId}-${cardef.cardNumber}`;
-            this.pool.set(fullId, cardef);
-        });
-    }
-
-    public size(): number {
-        return this.pool.size;
-    }
-
-    public lookup(cardId: CardId): Cardef | undefined {
-        return this.pool.get(cardId);
-    }
-
-    private loadFromFile(path: string): void {
-        try {
-            const tsvContents = readFileSync(path);
-            this.loadFromTsvString(tsvContents.toString('utf-8'));
-        } catch (e: unknown) {
-            console.log(e);
-            throw e;
-        }
-    }
-
-    private forceToString(s: string | undefined): string {
-        return s ?? '';
-    }
 }
